// src/lib-stock/models/ProductCategory.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductCategory = {
    id?: string;
    organizationId?: string;
    name?: string;
    description?: string;
    parentId?: string;
    code?: string;
    createdAt?: string;
    updatedAt?: string;
};